Added a lot of docs and did things here and there
